@@ -6,13 +6,8 @@ import sys
 import pathlib
 import shutil
 
+
 # Usage python main.py <infile> <outfolder>
-infile = pathlib.Path(sys.argv[1])
-outfolder = pathlib.Path(sys.argv[2])
-PIL.Image.MAX_IMAGE_PIXELS = 933120000
-
-
-
 
 def tile(filename, out_folder, d):
     name, ext = filename.stem, filename.suffix
@@ -23,19 +18,27 @@ def tile(filename, out_folder, d):
     for i, j in grid:
         box = (j, i, j+d, i+d)
         out = out_folder / f'{name}_{i}_{j}{ext}'
-        print(out)
+        print(out, flush=True)
         img.crop(box).save(out)
 
+def recreate_folder(f):
+    if os.path.isdir(f):
+            shutil.rmtree(f)
+    os.mkdir(f)
 
-if os.path.isdir(outfolder):
-    shutil.rmtree(outfolder)
-os.mkdir(outfolder)
 
+def main():
+    infile = pathlib.Path(sys.argv[1])
+    outfolder = pathlib.Path(sys.argv[2])
+    PIL.Image.MAX_IMAGE_PIXELS = 933120000
+    recreate_folder(outfolder)
 
-try:
-    tile(infile, outfolder, 400)
-    infile.unlink(missing_ok=True)
-    exit(0)
-except Exception as e:
-    print(e)
-    exit(1)
+    try:
+        tile(infile, outfolder, 400)
+        infile.unlink(missing_ok=True)
+    except Exception as e:
+        print(e, flush=True)
+        exit(1)
+
+if __name__ == '__main__':
+    main()
